Fix MSE bug and add rain-out to model
I realised that it was *actually* the liquid static energy and not moist
static energy which had been implemented in the cloud equations (and
also in Jeff's derivation). This fixes that problem.

Also introduces rain-out rate for removing rain-droplets and through
this introduces a rain-out length-scale (currently arbitrarily set to 100m)