Cloud equation mse gradient fixes
If I include consideration of the temperature dependency of the latent
heat then the thermodynamic equation turns out to be consistent with
what would be expected from modelling only microphysics and no
entrainment. This commit includes that change. Should also include the
temperature dependency of latent heat in the evaluation of latent heat.