Update parcel.py
Was playing around with the entrainment term to test for a different kind of parameterisation which requires the value for buoyancy in its calculation. I calculated buoyancy as: -g*(cloud density - environment density). This brought me to rho_c which, when generally has expected behaviour, but at some rare points has odd anomalies. Here is a section of the values: 
...
1.116959183193716
1.1039598870454252
1.097113145319669
0.7909797510557045
-0.00012469050121002718
nan
nan
1.114346875346376
1.1130510056341305
1.1065541081288586
...

Do you have any idea why this happens?